app/requirements.txt
envreader
dataclasses
psycopg2-binary
transitions==0.9.0
vk-api==11.9.9
vktools==1.0.6